+++ packages/terra-folder-tree/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
+* Added
+  * Created base `terra-folder-tree` component.
+
 ## 0.1.1-alpha.0 - (October 25, 2023)
 
 * Changed
-  * Minor dependency version bump
-
+  * Minor dependency version bump.